+++ web/documentserver-example/ruby/app/models/document_helper.rb
@@ -200,7 +200,7 @@ class DocumentHelper
       demo_name = (sample == 'true' ? 'sample.' : 'new.') + file_ext
       file_name = get_correct_name(demo_name, nil)  # get the correct file name if such a name already exists
 
-      src = Rails.root.join('public', 'assets', sample == 'true' ? 'sample' : 'new', demo_name)  # save sample document of a necessary extension to the storage directory
+      src = Rails.root.join('assets', 'document-templates', sample == 'true' ? 'sample' : 'new', demo_name)  # save sample document of a necessary extension to the storage directory
       dest = storage_path file_name, nil
 
       FileUtils.cp src, dest